requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
dashscope>=1.14.1
python-jose==3.3.0
django-filter==24.1
chromadb>=0.4.22
sentence-transformers>=2.2.2 